+++ src/components/layout/Navbar.tsx
@@ -2,8 +2,8 @@
 
 import React from "react";
 import Link from "next/link";
-import UserInfo from "@components/bars/UserInfo";
-import { Switcher } from "@components/dark-mode/Switcher";
+import UserInfo from "@/components/layout/UserInfo";
+import { Switcher } from "@/components/theme/Switcher";
 import { Button } from "@components/ui/button";
 import { useAuth } from "@/hooks/use-auth";
